working site & servlet prototype, action classes created, small fixes connPoll and dao

+++ src/main/java/com/epam/ot/products/Bullet.java
@@ -4,7 +4,7 @@ public class Bullet extends Product {
     private String caliber;
     private String type;
 
-    public Bullet(long id, String caliber, String name, String bulletType, Double price) {
+    public Bullet(int id, String caliber, String name, String bulletType, Double price) {
         super(id, name, price);
         this.caliber = caliber;
         this.type = bulletType;
